Optimize Preimage
This is a small improvement for GAP and a large improvement for Julia:
In Julia, `positions` is a list of *small* integers, so we avoid many
GMP operations. In general, the use of small integers is a problem for
correctness. Here, however, we have no chance of looping over
`AsList( phi )` anyway if its length is not a small integer in the first
place.